refactor: remove unused legacy model & move DatasetSearchConfig
