--- gms/api/src/main/pegasus/com/linkedin/dataset/LineageType.pdsc
@@ -1,18 +1,0 @@
-{
-  "type": "enum",
-  "name": "LineageType",
-  "namespace": "com.linkedin.dataset",
-  "doc": "The various types of lineage supported",
-  "symbols": [
-    "DIRECT_COPY",
-    "TRANSFORMED",
-    "VIEW",
-    "MANAGE"
-  ],
-  "symbolDocs": {
-    "DIRECT_COPY": "Direct copy without modification",
-    "TRANSFORMED": "Transformed data with modification (format or content change)",
-    "VIEW": "Represent lineage data like database view. i.e. HIVE use VIEW type to show HDFS stored datasets",
-    "MANAGE": "Represents a higher level of source dataset (e.g. HIVE), managing a lower level of destination dataset (e.g. HDFS)"
-  }
-}
